+++ PlanBuild/ModCompat/PatcherGizmo.cs
@@ -9,15 +9,10 @@ namespace PlanBuild.ModCompat
 {
     internal class PatcherGizmo
     {
-        private static Transform gizmoRoot;
-        private static float snapAngle;
-
         [HarmonyPatch(typeof(GizmoReloaded.Plugin), "UpdatePlacement")]
         [HarmonyPrefix]
         private static bool GizmoPlugin_UpdatePlacement_Prefix(Transform ___gizmoRoot,  float ___snapAngle)
         {
-            gizmoRoot = ___gizmoRoot;
-            snapAngle = ___snapAngle;
             if (Player.m_localPlayer && Player.m_localPlayer.m_buildPieces && Player.m_localPlayer.m_placementGhost &&
                 Player.m_localPlayer.m_buildPieces.name.StartsWith(BlueprintAssets.PieceTableName) &&
                 !Player.m_localPlayer.m_placementGhost.name.StartsWith(Blueprint.PieceBlueprintName))
@@ -51,23 +46,13 @@ namespace PlanBuild.ModCompat
         {
             if (axis == "Y" && Player.m_localPlayer && Player.m_localPlayer.m_buildPieces &&
              Player.m_localPlayer.m_buildPieces.name.StartsWith(BlueprintAssets.PieceTableName) &&
-             ZInput.GetButton(BlueprintConfig.CameraModifierButton.Name))
+             (ZInput.GetButton(BlueprintConfig.CameraModifierButton.Name) ||
+              ZInput.GetButton(BlueprintConfig.DeleteModifierButton.Name) ||
+              ZInput.GetButton(BlueprintConfig.RadiusModifierButton.Name)))
             {
                 return false;
             }
             return true;
         }
-
-        [HarmonyPatch(typeof(ToolComponentBase), "UndoRotation")]
-        [HarmonyPrefix]
-        private static bool ToolComponentBase_UndoRotation_Prefix(Player player, float scrollWheel)
-        {
-            if(gizmoRoot)
-            {
-                gizmoRoot.rotation *= Quaternion.AngleAxis(-Mathf.Sign(scrollWheel) * snapAngle, Vector3.up);
-            }
-            return false;
-        }
-
     }
 }